+++ contracts/SafeMath.sol
@@ -1,38 +1,39 @@
 pragma solidity 0.5.11;
 
+
 /**
  * Math operations with safety checks
  */
 library SafeMath {
-    function mul(uint a, uint b) internal pure returns (uint) {
-        uint c = a * b;
+    function mul(uint256 a, uint256 b) internal pure returns (uint256) {
+        uint256 c = a * b;
         require(a == 0 || c / a == b);
         return c;
     }
 
-    function div(uint a, uint b) internal pure returns (uint) {
+    function div(uint256 a, uint256 b) internal pure returns (uint256) {
         require(b > 0);
-        uint c = a / b;
-        require(a == b * c + a % b);
+        uint256 c = a / b;
+        require(a == b * c + (a % b));
         return c;
     }
 
-    function sub(uint a, uint b) internal pure returns (uint) {
+    function sub(uint256 a, uint256 b) internal pure returns (uint256) {
         require(b <= a);
         return a - b;
     }
 
-    function add(uint a, uint b) internal pure returns (uint) {
-        uint c = a + b;
+    function add(uint256 a, uint256 b) internal pure returns (uint256) {
+        uint256 c = a + b;
         require(c >= a);
         return c;
     }
 
-    function max(uint a, uint b) internal pure returns (uint) {
+    function max(uint256 a, uint256 b) internal pure returns (uint256) {
         return a >= b ? a : b;
     }
 
-    function min(uint a, uint b) internal pure returns (uint) {
+    function min(uint256 a, uint256 b) internal pure returns (uint256) {
         return a < b ? a : b;
     }
 }
